Updated to introduce type synonyms (aliases)
This brings several potential code-quality benefits:

Defining the type synonym in one place facilitates changing that type later.
Using a custom name for a type reduces the chance of confusion with other aspects of the code that share the same implementation type (e.g. integers may also be used to record the size of the dictionary);
Using language from the domain being implemented improves the readability of the code. (When discussing the theory of dictionaries we talk about keys and items, therefore it is helpful if the code also refers to keys and items, rather than integers and strings.)